+++ CHANGELOG.rst
@@ -3,6 +3,12 @@
 Change Log
 ==========
 
+3.9.0
+=====
+* Added functionality to run sample identity checks
+* Add `dry_run` option to `archive_unaligned_reads` command
+
+
 3.8.0
 =====
 * Added possibility to overwrite exsting QCs when running a standalone QC workflow.